fix(observability): redact finalization stream tokens from performance logs
